Modify detail design in WhoAreYouPage
Add the gradient background
Add shadow below the question text box
Seperate `FavoriteIcon` from `AnswerButtons`
