+++ index.html
@@ -13,6 +13,7 @@
       margin: 0;
       padding: 0;
       box-sizing: border-box;
+      word-break: keep-all;
     }
     body {
       margin: 0 auto;
@@ -22,6 +23,9 @@
       font-size: 16px;
       line-height: 1.4;
       background: #DDD;
+      background-repeat: no-repeat;
+      background-attachment: fixed;
+      background-image: linear-gradient(50deg, rgba(252,238,118,1), rgba(216,174,204,1));
     }
     main {
       background: #FFF;
