+++ src/components/who-are-you/TopBar.jsx
@@ -15,7 +15,7 @@ const Container = styled.div({
   maxWidth: '500px',
   height: TOP_BAR_HEIGHT,
   background: '#FFF',
-  boxShadow: '0 3px 3px -3px rgba(0,0,0,.2)',
+  boxShadow: '0 4px 4px -4px rgba(0,0,0,.2)',
   '& img': {
     height: '3em',
     position: 'absolute',
